requirements.txt
alembic==1.13.1
apispec==6.4.0
blinker==1.7.0
click==8.1.7
flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
flask-smorest==0.42.3
flask-sqlalchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
packaging==23.2
passlib==1.7.4
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing-extensions==4.9.0
webargs==8.4.0
werkzeug==3.0.1
zipp==3.17.0
